# server/config/swagger/kibopush.neweditor.yaml
...
swagger: '2.0'
...
info:
  version: v1
  title: KiboPush API
  description: |
    Welcome to the KiboPush API documentation! Our API provides flexible tools that can be used in the creation of new, outstanding projects.
    Please note that this documentation refers to the API 1.0 version. API 2.0 version is currently under construction and will be released soon.

    Build something great!

    ## Basic Usage
    All KiboPush API requests start with `https://app.kibopush.com/api/`

    The next segment of the URI path depends on the type of request. For example:

    `https://app.kibopush.com/api/users/`

    to get information of logged in user.

    ## Authentication
    In order for client application to call our API, he/she would have to attach the following information to the http headers:

      * app_id (the API key)
      * app_secret (the API key secret)


    ## Error Handling
    Errors are returned using standard HTTP error code syntax. In general, codes in the 2xx range indicate success, codes in the 4xx range indicate an error (bad or missing parameters, not authenticated etc.), and codes in the 5xx range indicate an error with KiboPush’s servers. Any additional info is included in the body of the return call, JSON-formatted.

  termsOfService: http://www.kibopush.com/privacy-policy

################################################################################
#                  Host, Base Path, Schemes and Content Types                  #
...
host: app.kibopush.com
basePath: /api
schemes:
...
produces:
...
consumes:
...
securityDefinitions:
  apiKey:
    type: apiKey
    in: header
    name: app_id
  apiSecret:
    type: apiKey
    in: header
    name: app_secret
security:
  - apiKey: []
    apiSecret: []

################################################################################
#                                           Paths                              #
################################################################################
paths:
...
  /growthtools/upload/:
    post:
      description: uploads the csv file containing customers' name and phone numbers
      consumes:
          - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: file
          in: formData
          description: The file to upload
          type: file
          required: true
        - name: text
          in: formData
          description: The message text
          type: string
          required: true
        - name: pageId
          in: formData
          description: The object id of the page
          type: string
          required: true
        - name: _id
          in: formData
          description: The string id of the page
          type: string
          required: true
        - name: nameColumn
          in: formData
          description: list of names
          type: string
          required: true
        - name: phoneColumn
          in: formData
          description: list of phone numbers
          type: string
          required: true

      responses:
        '201':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              description:
                type: string
                description: Contacts were invited to your messenger
        '404':
          description: failed
          schema:
             $ref: '#/definitions/Response404'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
  /growthtools/sendNumbers/:
    post:
      description: Sends a message to the phone numbers entered
      consumes:
          - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: growthtool
          in: body
          description: Message details
          required: true
          schema:
            type: object
            properties:
              numbers:
                type: array
                items:
                  type: string
              pageId:
                type: string
                description: Object Id of the page
              _id:
                type: string
                description: String Id of the page
              text:
                type: string
                description: text of the message
      responses:
        '201':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              description:
                type: string
                description: Contacts were invited to your messenger
        '404':
          description: failed
          schema:
             $ref: '#/definitions/Response404'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
  /growthtools/pendingSubscription/:name/:
    get:
      description: Retrieves phone of a particular file that haven't subscribed yet
      consumes:
          - application/json
      produces:
        - application/json
      responses:
        '201':
          description: OK
          schema:
            type: object
            properties:
              status:
               type: string
               description: success
              payload:
                type: array
                items:
                  $ref: '#/definitions/PhoneNumbers'
        '404':
          description: failed
          schema:
             $ref: '#/definitions/Response404'
        '500':
          description: failed
          schema:
             $ref: '#/definitions/Response500'
...
definitions:
...
  PhoneNumbers:
    type: object
    description: The table stores the phone numbers of the customers
    properties:
      _id:
        type: string
        description: unique id of the record
        readOnly: true
      pageId:
        type: string
        description: the id of the page to set webhook for
      companyId:
        type: string
        description: company id of the company that the user belongs to
        readOnly: true
      userId:
        type: string
        description: user id of the user
        readOnly: true
      name:
          type: string
          description: Name of the customer
      number:
        type: string
        description:  Number of the customer
      hasSubscribed:
        type: boolean
        default: false
        description: If the user has subscribed to the page or not
      fileName:
        type: array
        items:
          type: string
...
  Response500:
    type: object
    properties:
      status:
        type: string
        description: failed
      description:
        type: string
        description: Internal server error
...
  Response404:
    type: object
    properties:
      status:
        type: string
        description: failed
      description:
        type: string
        description: Record not found
...